@@ -37,8 +37,8 @@
   "dependencies": {
     "date-fns": "^3.6.0",
     "mitt": "^3.0.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "styled-components": "^6.1.8"
   }
 }
