@@ -1,4 +1,5 @@
-import ReactDOM from "react-dom";
+import { createRoot } from 'react-dom/client';
+
 import React from "react";
 import './index.css';
 import Advanced from "./components/AdvancedReact/Advanced";
@@ -8,4 +9,5 @@ function App() {
     return <Advanced />;
 }
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const root = createRoot(document.getElementById("root"))
+root.render(<App />);
