@@ -1,10 +1,11 @@
 import React from 'react';
 import CallbackMemo from './CallbackMemo';
+import ImperativeHandle from './ImperativeHandle';
 import Refs from './Refs';
 
 function Advanced() {
   return (
-    <Refs/>
+    <ImperativeHandle/>
   )
 }
 
